chore: down arrow added, helper functions moved into service.ts
