+++ common/custom-icons/customIcons.ts
@@ -81,5 +81,5 @@ export const RightChevron = createIcon({
 export const DownChevron = createIcon({
   displayName: "DownChevron",
   viewBox: "0 0 19 20",
-  d: "M7.893 6.413a.594.594 0 01.84 0l3.166 3.167a.594.594 0 010 .84l-3.167 3.166a.594.594 0 11-.84-.84L10.64 10 7.893 7.253a.594.594 0 010-.84z",
+  d: "M13.087 7.601a.594.594 0 010 .84L9.92 11.607a.594.594 0 01-.84 0L5.914 8.441a.594.594 0 01.84-.84L9.5 10.348 12.247 7.6a.594.594 0 01.84 0z",
 });
